requirements.txt
streamlit==1.35.0
scikit-learn==1.5.0
google-generativeai==0.5.4
python-docx==1.1.0
pymupdf==1.24.2
python-dotenv==1.0.1